Display progress of loading weights in terminal
This change fixes a longstanding UX issue with llama.cpp which is that
the first inference evaluation might take literally minutes to page-in
memory from disk that was mmap()'d. This change prefaults memory pages
manually and prints the percentage loaded to an ephemeral line in your
terminal. That way, it will not appear as if your llama.cpp has frozen