Introduce --silent-prompt flag to main
This makes it much easier to use llamafile-main in a shell script that
provides instructions to Mistral. For example, if I want to summarize a
news article, I can say:

  o//llama.cpp/main/main -m ~/weights/mistral-7b-instruct-v0.1.Q4_K_M.gguf \
    -c 8000 -p "$(echo $'[INST]Summarize the following article:\n'; links \
    -codepage utf-8 -force-html -width 72 -dump \
    https://www.troyhunt.com/a-decade-of-have-i-been-pwned/; echo \
    $'\n[/INST]\n')" --temp 0 --silent-prompt 2>/dev/null

And the only thing that'll be printed in my terminal is the summary.
This issue was reported in ggerganov/llama.cpp#2664 but it appears it
was closed as a wontfix issue.