Have makefile download cosmocc automatically
This change ensures builds will be reproducible now and forevermore
because it pins our only dependency (which is cosmocc) using shasum
checked download at a specific version number.

This change ensures our development environment is turn-key. If the
system has the POSIX shell and a modern version of GNU Make then it
should be guaranteed to work. I've tested this on Linux, MacOS, and
Windows which needs to have the Cosmopolitan bash shell from Cosmos