Slightly speed up LLaVA runtime dispatch on Intel
On Intel Core i9-9900 w/ 2667 MT/s RAM this change reduces the time it
takes to analyze a photo of lemurs from 46 seconds to 42 seconds. This
means we're almost back at parity with llama.cpp, which takes 40 secs.
Chances are the difference is more noticeable on systems w/ faster RAM