Change README to list Mixtral Q5 (instead of Q3)
The Q3 weights hallucinate far too much for me to feel comfortable
listing them here. It's also been reported on the llama.cpp issues
tracker, that the Q6 weights have worse perplexity scores than Q5.
The consensus online is that Q5 or better yet Q8 are the way to go