Read and write column-major matrices better (#164)
* read matrix A better

parallelizing with respect to BK improves speed further, because of how
the matrix A is laid out in memory. This performance improvement takes
advantage of the CUBLAS_OP_T for A.

* write back to global memory a bit better

this gives like 15 tokens/s boost in prompt eval time, because now there
is less competition between the threads when writing back to global
memory. (there are less threads overall, because BN is usually less than
BM, and we force fewer threads when writing back to global memory).

---------

Co-authored-by: Gautham <ahgamu@gmail.com>