dupes in sharedmem to avoid bank conflicts
when reading As, there will be BN/TN threads that have the same ti, and
all of them will want to read the same value.

when reading Bs, there will be BM/TM threads that have the same tj, and
all of them will want to read the same value.

so now we have dupes so each of those can read nearby values at least.
this can likely be arranged more optimally so that every SMEM read is
coalesced.