More conservative strong/em markdown matcher
Both matchers are now constrained so that they will only be transformed
if the left-hand characters are the start of the text or preceeded by
whitespace, and the right-hand characters are the end of the text or
followed by whitespace or a punctuation mark.

This misses some cases, e.g. emphasis inside of parantheticals, but has
better behavior with e.g. underscores in function names outside of code
blocks.

Fixes #317.