Synchronize after every op
The previous commit loosened the synchronization barriers too much, and
it caused some rare brain damage when using quants. With this change,
we're now only using 3-4x fewer sychronization barriers than before.

A new --trap flag has been introduced by this change, which enables the
runtime to detect the precise moment a NaN appears anywhere in the code
and then aborts with both a C++ backtrace and a GGML graph trace.

Some jankiness with pledge() and /dev/tty is also fixed by this change.