Make GGML asynchronously cancelable
It's now possible to instantly kill jobs. Even if a thread is stuck in a
very long matmul operation, an asynchronous signal will be sent, to make
pthread_exit() run immediately. Doing this does not leak memory. When it
happens, a 503 Service Unavailable response is sent to the client, so it
knows to try again with exponential backoff. It'll help with reliability
in the event of DDOS. It'll also help with prioritization of tasks. With
this capability, we could have an HTTP header where clients volunteer to
be preemptable, which could be useful for jobs running in the background