Disable GGML_CUDA_FORCE_MMQ
We enabled this option earlier because it goes faster than tinyBLAS. It
is however enormous. It's grown so much in size over the past month, it
is going to cause the llamafile executable to increase from 30mb to 219
megabytes which is unbelievable. Unfortunately, disabling this will not
save us from this bloat. Despite adding so much bloat, the new MMQ code
appears to have dropped support for my five year old GeForce RTX 2080Ti
which is probably an unintended bug but we need this to unblock release