Check-in llama matmul experiment
This code is the result of a collaboration between Aman Salykov and
Justine Tunney. Aman managed to figure out a simple, elegant way to
implement the BLIS paper, which is something I spent months trying.
However I could never get it to work nearly as well as this code.

Both BLIS and MKL design their microkernels to use the broadcast
instruction to load elements from the B matrix. This doesn't reduce the
number of memory load operations compared to tinyBLAS, but it does
reduce the number of bytes that need to be transferred. For example,
with BF16, that should mean we can just load a 16-byte value whenever we
need to fetch data from B rather than needing to pull in a whole vector.

In order for the broadcast trick to work, you have parallelize across
the K dimension, which is tricky. That's because you need to load tiles
from both the A and B matrices before calling the microkernel, and they
need to be rearranged, on "packed", in a special way.

What impresses me about Aman's technique is that his code appears to be
capable of parallelizing dot products on the K dimension across threads
without using atomics or locks. It only needs two specially placed CUDA
style syncthreads() directives, which impose an order on memory ops.

See https://salykova.github.io/matmul-cpu
See https://github.com/salykova/matmul.c