Always favor fp16 arithmetic in tinyBLAS
It was assumed earlier that upcasting would help precision. However this
wasn't the case, according to levenshtein distance in whisperfile output
which tells me this change makes things objectively better in noticeable
ways. So we now avoid the fp16 conversion, when the ISA is available. It
should be perfectly safe and accurate, even for large sums, since we now
have the ruler reduction divide and conquer approach, in tinyBLAS::gemm.