Introduce /upload and /forget commands to chatbot
The /upload command allows you to share an image or a text file with the
assistant. In the case of images, it'll be printed in the terminal, then
fed through LLaVA so that --mmproj vision models can analyze it for you.

The /forget command may be used to erase the oldest chat messages from a
context window. This is useful for salvaging a conversation that got too
long. When running out of context, the chatbot will no not exit anymore,
but instead return control to the REPL so you can free up the resources.