Avoid streaming incomplete UTF-8 characters
Some characters, like the chinese fù is sometimes returned as two tokens,
as "\u00e8\u00b5" and "\u008b" in this case.

This is also depends on the model, but when it happens, for example
with DeepSeek R1, we have to wait for the character to be complete and
send it only then.

This resolves #722 and #646